Organize msg components & move componenets into related folders
