+++ backend/src/controllers/channelController/channelGetMsgs.ts
@@ -24,7 +24,7 @@ export default async function channelGetMsgs(
       throw new Error("Channel not found");
     }
 
-    return res.status(200).json({ channelMsgs });
+    return res.status(200).json({ messages: channelMsgs });
   } catch (error) {
     return next(error);
   }
