+++ backend/src/controllers/dmController/dmGroupGetMsgs.ts
@@ -19,7 +19,7 @@ export default async function dmGetMsgs(
 
     const dmMsgs = await getDmMsgs(groupId);
 
-    return res.status(200).json({ dmMsgs });
+    return res.status(200).json({ messages: dmMsgs });
   } catch (error) {
     next(error);
   }
