+++ frontend/src/components/modals/JoinSpaceModal.tsx
