+++ frontend/src/components/user/UserSpaceList.tsx
@@ -1,14 +1,12 @@
 import { useNavigate, useParams } from "react-router-dom";
-import { UserSpace } from "../types/types";
+import { UserSpace } from "../../types/types";
 
-export default function SpaceList({ spaces }: { spaces: UserSpace[] }) {
+export default function UserSpaceList({ spaces }: { spaces: UserSpace[] }) {
   const { spaceId } = useParams<{ spaceId: string }>();
   const navigate = useNavigate();
   const handleSpaceClick = (spaceId: string, defaultChannel?: string) => {
     navigate(`/channels/${spaceId}/${defaultChannel}`);
   };
-
-  if (spaces?.length === 0) return <div>Loading</div>;
 
   return (
     <div className="space-list">
